tidy code formatting

+++ tests/testthat/helper-csv.R
@@ -4,7 +4,10 @@ test_csv_file <-
 set.seed(0)
 
 test_csv_data <-
-  data_frame(foo = "bar", baz = as.integer(1:3), bap = c(147, NA, 1 / 128))
+  data_frame(
+    foo = "bar",
+    baz = as.numeric(1:3),
+    bap = c(147, NA, 1 / 128))
 
 # Use only the base::write.csv() function
 # (not write_csv() from this package or from the readr package)
