+++ tests/testthat/test-write_csv.R
@@ -40,9 +40,10 @@ test_that("finite signif read back", {
 
   expect_equal(
     read_back,
-    data_frame(foo = as.character("bar"),
-               baz = as.integer(1:3),
-               bap = c(147, NA, signif(1 / 128, 4))))
+    data_frame(
+      foo = as.character("bar"),
+      baz = as.numeric(1:3),
+      bap = c(147, NA, signif(1 / 128, 4))))
 
   csv_lines <- readLines(tmpfn)
   expect_equal(
@@ -61,9 +62,10 @@ test_that("Inf signif read back", {
     "use `signif`")
 
   read_back <- read_csv(tmpfn)
-  expected <- data_frame(foo = as.character("bar"),
-                         baz = as.integer(1:3),
-                         bap = c(147, NA, 0.0078125))
+  expected <- data_frame(
+    foo = as.character("bar"),
+    baz = as.numeric(1:3),
+    bap = c(147, NA, 0.0078125))
 
   expect_equivalent(
     read_back,
@@ -85,9 +87,10 @@ test_that("default signif", {
 
   write_csv(test_csv_data, verbose = TRUE, tmpfn)
   read_back <- read_csv(tmpfn)
-  expected <- data_frame(foo = as.character("bar"),
-                         baz = as.integer(1:3),
-                         bap = c(147, NA, signif(1 / 128, getOption("digits"))))
+  expected <- data_frame(
+    foo = as.character("bar"),
+    baz = as.numeric(1:3),
+    bap = c(147, NA, signif(1 / 128, getOption("digits"))))
   expect_equal(read_back, expected)
 
   csv_lines <- readLines(tmpfn)
